ref(xcode-ide): Simplify bridge tool handler wiring
Extract shared bridge handler lookup so xcode-ide tools follow one path\nfor bridge availability and fallback behavior.\n\nExtract shared remote tool serialization used by both manager and\nstandalone bridge implementations to reduce duplication and drift.\n\nUpdate bridge tool tests/mocks to align with shared exports and keep\ncoverage around standalone fallback behavior.
