+++ src/mcp/tools/xcode-ide/xcode_tools_bridge_sync.ts
@@ -1,14 +1,8 @@
 import type { ToolResponse } from '../../../types/common.ts';
-import { getServer } from '../../../server/server-state.ts';
-import { getXcodeToolsBridgeToolHandler } from '../../../integrations/xcode-tools-bridge/index.ts';
-import { createErrorResponse } from '../../../utils/responses/index.ts';
+import { withBridgeToolHandler } from './shared.ts';
 
 export const schema = {};
 
 export const handler = async (): Promise<ToolResponse> => {
-  const bridge = getXcodeToolsBridgeToolHandler(getServer());
-  if (!bridge) {
-    return createErrorResponse('Bridge unavailable', 'Unable to initialize xcode tools bridge');
-  }
-  return bridge.syncTool();
+  return withBridgeToolHandler(async (bridge) => bridge.syncTool());
 };
